feat: userID 타입 수정 및 global exception 수정

See merge request s14-webmobile1-sub1/S14P11D108!63

+++ Backend/src/main/java/com/ssafy/d108/backend/global/exception/OrderNotFoundException.java
@@ -11,7 +11,7 @@ public class OrderNotFoundException extends BusinessException {
         super(ErrorCode.ORDER_NOT_FOUND);
     }
 
-    public OrderNotFoundException(Long orderId) {
+    public OrderNotFoundException(Integer orderId) {
         super(ErrorCode.ORDER_NOT_FOUND, "주문 내역을 찾을 수 없습니다. ID: " + orderId);
     }
 }
